Add FileBot support
FileBot generally has better support:
  * Compressed torrents
  * Identifying and renaming movies
  * Picking up subtitles

However, FileBot has a difficult time identifying some foreign movies, which Bobarr
can find because Bobarr is aware of what media the user is trying to download
in the first place.

This commit tries to get the best of both worlds:
  1. Let Bobarr make directories to give FileBot a hint.
  2. Let FileBot do the rest, and output to a final directory.

If FileBot outputs to the same `MOVIES_FOLDER` and `TV_SHOWS_FOLDER`
directories that Bobarr expects, Bobarr will eventually pickup and associate
the media files with the show objects when it periodically scans.

Bobarr makes a directory in a `stage` directory that is named the same
name as the torrent we are trying to add add.

    `Some.File.I.Wanted.S01.E42/Some File I Wanted/Season 01/`

When the torrent is done, Transmission will run a script on the
completed torrent, which will look to the stage directory first, to see
if there are any directories named the same as the torrent that just
finished.  If so, it will use that directory structure when sending
files to the FileBot queue.  If no staged files are found, it will send
the files as-is to FileBot.

This also allows users to add torrents to Transmission directly, and
have them show up in Bobarr.